Add initial implementation of Gastby Wordpress Gutenberg plugin
This plugin is added to the root of the repo to be able to added to packagist later on.
Folder structure is inspired from gutenberg git repo.

Plugin is also added to default example docker backend.
Dependency on the wp-graphql-gutenberg is removed.

Comments to the php code will be added later.